requirements.txt
requests
pyspark
torch
torchvision
boto3
numpy
pillow